test(integration): consolidate deleteAppMap and openAsJson into one Electron run
Both suites use the default project-a workspace, identical setup, and exercise the
custom-editor context commands. Merge them into a single file so they share one
Electron instance instead of booting two. Factor the shared appmap path and
tab-lookup helpers.

Integration test files: 27 -> 25.
